ALl test case pass

+++ Algorithm/DynamicProgramming/IntroductionDynamicProgramming/3:XSquareAndTwoArray/Solution.java
@@ -7,22 +7,22 @@ class TestClass {
       
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringBuilder sb = new StringBuilder();
-        int[] a = new int[111100];
-        int[] b = new int[111100];
-        int[] a1 = new int[111100];
-        int[] b1 = new int[111100];
+        long[] a = new long[111100];
+        long[] b = new long[111100];
+        long[] a1 = new long[111100];
+        long[] b1 = new long[111100];
         String name[] = br.readLine().split(" ");                // Reading input from STDIN
-        int n = Integer.parseInt(name[0]);
-        int q = Integer.parseInt(name[1]);
+        long n = Long.parseLong(name[0]);
+        long q = Long.parseLong(name[1]);
         String[] arr = br.readLine().split(" ");
         String[] brr = br.readLine().split(" ");
         for(int i = 1;i <= n;i++)
         {
-             a[i] = Integer.parseInt(arr[i-1]);
+             a[i] = Long.parseLong(arr[i-1]);
         }
         for(int i = 1 ; i <=n; i++)
         {
-            b[i] = Integer.parseInt(brr[i-1]);
+            b[i] = Long.parseLong(brr[i-1]);
         }
         for(int i = 1 ; i <= n; i++)
         {
